backend/requirements/base.txt
Django>=5.1,<5.2
django-tenants>=3.7
django-unfold>=0.81
djangorestframework>=3.15
django-cors-headers>=4.3
django-allauth>=65.0
dj-rest-auth[with_social]>=7.0
djangorestframework-simplejwt>=5.3
django-filter>=24.0
psycopg2-binary>=2.9
Pillow>=10.0
whitenoise>=6.6
gunicorn>=22.0
python-dotenv>=1.0
